+++ frontend/composables/useAuthUser.ts
@@ -15,14 +15,22 @@ export const useAuthUser = () => {
             return;
         }
 
-        const { $auth } = useNuxtApp();
+        // ★★★ 修正箇所: ここから変更 ★★★
 
-        // onAuthStateChanged: ログイン状態が変化したときに呼び出される
-        onAuthStateChanged($auth, (currentUser) => {
-            user.value = currentUser;
-            isAuthReady.value = true; // 初回コールで初期化完了
-            console.log('認証状態が変化しました:', currentUser ? `UID: ${currentUser.uid}` : 'ログアウト');
-        });
+        // Nuxtのサーバーサイドレンダリング(SSR)時にFirebase Authコードを実行するのを防ぐ
+        if (process.client) {
+            const { $auth } = useNuxtApp();
+
+            // onAuthStateChanged: ログイン状態が変化したときに呼び出される
+            onAuthStateChanged($auth, (currentUser) => {
+                user.value = currentUser;
+                isAuthReady.value = true; // 初回コールで初期化完了
+                console.log('認証状態が変化しました:', currentUser ? `UID: ${currentUser.uid}` : 'ログアウト');
+            });
+        } else {
+            // サーバーサイドでは、とりあえず準備完了フラグを立てて、レンダリングを続行させる
+            isAuthReady.value = true;
+        }
     };
 
     init();
